Corregido ejercicio 1

+++ Tema 2/Ejercicio01.java
@@ -2,15 +2,15 @@
 public class Ejercicio01 {
 	
 	public static void main (String[] args) {
-		int x=144;
-		int y=999;
+		int x = 144;
+		int y = 999;
 		
 		System.out.println("x=" + x);
 		System.out.println("y=" + y);
-		System.out.println("x+y=" + (x+y));
-		System.out.println("x-y=" + (x-y));
-		System.out.println("x/y=" + (x/y));
-		System.out.println("x*y=" + (x*y));
+		System.out.println("x+y=" + (x + y));
+		System.out.println("x-y=" + (x - y));
+		System.out.println("x/y=" + ((double)x / (double)y));
+		System.out.println("x*y=" + (x * y));
 	}
 }
